Update keyword dfa to remove banned words Also created banned word dfa's to abort early
