+++ scripts/banned_joos_keywords.txt
@@ -1,0 +1,18 @@
+catch
+continue
+default
+do
+double
+finally
+float
+long
+switch
+synchronized
+private
+transient
+super
+volatile
+try
+throw
+strictfp
+throws
