+++ scripts/joos_keywords.txt
@@ -4,20 +4,13 @@ boolean
 break
 byte
 case
-catch
 char
 class
 const
-continue
-default
-do
-double
 else
 enum
 extends
 final
-finally
-float
 for
 if
 goto
@@ -26,25 +19,14 @@ import
 instanceof
 int
 interface
-long
 native
 new
 package
-private
 protected
 public
 return
 short
 static
-strictfp
-super
-switch
-synchronized
 this
-throw
-throws
-transient
-try
 void
-volatile
 while
